Codecs renamed
